xen/riscv: select EARLY_PRINTK config
It might happen that randconfig will decide not to enable
early_printk() functionality and it will break CI test
as it is expected to see "All is set up".

There is no any other way to see the messages except usage
of early_printk functionality right now as Xen common code is not
available at the momemnt so common's printk() isn't available and the
current implementation of printk() uses early_printk()).

Signed-off-by: Oleksii Kurochko <oleksii.kurochko@gmail.com>