xen/riscv: switch to vgic functions instead of plic functions
These changes can be incorporated with the update of the PLIC code
to align with the generic interrupt controller (GIC) abstraction.
However, my expectation is that this commit will be automatically
dropped during the upstreaming of the GIC.

This upstreaming is anticipated to occur earlier than the introduction
of the `struct domain` or domain creation.

Signed-off-by: Oleksii Kurochko <oleksii.kurochko@gmail.com>