xen/riscv: do proper guest context save and restore
If an interrupt happens because of guest domain it is necessary
to save/restore CSR registers; otherwise it possible that some
guest events will be missed and undefined behaviour will be in
guest domain.

For example, it would be an issue with jiffies in guest domain
when SSTC extenstion isn't available, then timer will not trigger
an interrupt to guest domain, so jiffies won't be updated.

Signed-off-by: Oleksii Kurochko <oleksii.kurochko@gmail.com>